requirements.txt
cssutils==1.0.2
cx-Freeze==5.1.1
lxml==4.2.1
Pillow==5.1.0
psutil==5.4.5
PyHyphen==3.0.1
PyQt5==5.10.1
python-slugify==1.2.5
